Replace toInternalURL() with toURL() for Android
When the plugin is used in newer versions of Android, it always uses toInternalURL() instead of toURL(). When using toInternalURL() for Android, the plugin is unable to download and save or retrieve any cache files since the url provided by toInternalURL() is invalid.

By adding the above checking, this will solve the issue completely. Feel free to refactor the code if this fix doesn't apply to all scenarios tested.